profile: generate 20.90 SDK
Thanks to @unusedbytes for the following info:

"There's an error in the 20.88 spreadsheet, that in the dive_settings
message, the heart_rate_source_type field is not enabled in the example
product configuration (column P), but the heart_rate_source field which
refers to it is. This causes errors in generation, so to fix it, the
heart_rate_source_type field was enabled before running this generation.

The profile now includes a new exercise_category field, meaning an
update to the reader test golden values is required."