file_types: Move DeviceInfo to file bodies
In the SDK specification, it indicates that some files can have multiple
DeviceInfo messages, e.g. to track battery voltage over time, or for
multiple devices.

To support this, remove File.DeviceInfo and replace it with slices of
DeviceInfoMsgs in the individual file types which support it. According
to SDK 21.32.00 that's ActivityFile, WeightFile, BloodPressureFile,
MontoringAFile and MonitoringBFile.

Also update the fingerprints, either because File.DeviceInfo is gone, or
because new DeviceInfoMsgs are added.

Note: This is a breaking change! File.DeviceInfo no longer exists.

Fixes: #46