reader: tolerate strings without NULL terminator
At least one major manufacturer is producing strings in
their FIT file without a NULL terminator.

This change makes it so that the entirety of the
contents of the data field can be used as the string,
instead of assuming the last part will be a NULL
terminator.

A testdata file has also been added which includes
strings without a NULL terminator.